Refactored resolvers
